Split DataItem
Define a new inherited class: TypeNeutralData
This will allow in the near future to switch to Poco::Any usage, and use typeid instead of custom DataType
